requirements.txt
ansible==4.7.0
ansible-compat==0.5.0
ansible-core==2.11.6
ansible-lint==5.2.0
arrow==1.2.0
attrs==21.2.0
bcrypt==3.2.0
binaryornot==0.4.4
bracex==2.1.1
Cerberus==1.3.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-help-colors==0.9.1
colorama==0.4.4
commonmark==0.9.1
cookiecutter==1.7.3
cryptography==35.0.0
distro==1.6.0
docker==5.0.3
enrich==1.2.6
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.2
jinja2-time==0.2.0
MarkupSafe==2.0.1
molecule==3.5.2
molecule-docker==1.0.2
packaging==21.0
paramiko==2.8.0
pathspec==0.9.0
pluggy==1.0.0
poyo==0.5.0
py==1.10.0
pycparser==2.20
Pygments==2.10.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.5
pytest-testinfra==6.4.0
python-dateutil==2.8.2
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
resolvelib==0.5.4
rich==10.12.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
selinux==0.2.1
six==1.16.0
subprocess-tee==0.3.5
tenacity==8.0.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
wcmatch==8.2
websocket-client==1.2.1
yamllint==1.26.3
